add id size checks

+++ backend/api/schemas/delete.go
@@ -33,6 +33,9 @@ func (c *Controller) apiDelete(ctx *fiber.Ctx) error {
 	if err != nil {
 		return err
 	}
+	if len(body.IDs) == 0 {
+		return fiber.ErrBadRequest
+	}
 
 	err = c.schemas.DeleteMany(ctx.Context(), repository.IDsIn(body.IDs))
 	if err != nil {
